+++ Plugins/Core/Source/NexusCore/Public/NMultiplayerLibrary.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "NMultiplayerMinimal.h"
+#include "NCoreMinimal.h"
 #include "NMultiplayerUtils.h"
 #include "UObject/ObjectMacros.h"
 #include "Kismet/BlueprintFunctionLibrary.h"
@@ -15,10 +15,10 @@ class APlayerState;
 
 /**
  * A handful of methods meant to support the building logic that works in multiplayer scenarios.
- * @see <a href="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/">UNMultiplayerLibrary</a>
+ * @see <a href="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/">UNMultiplayerLibrary</a>
  */
 UCLASS(ClassGroup = "NEXUS", DisplayName = "NEXUS | Multiplayer Library")
-class NEXUSMULTIPLAYER_API UNMultiplayerLibrary : public UBlueprintFunctionLibrary
+class NEXUSCORE_API UNMultiplayerLibrary : public UBlueprintFunctionLibrary
 {
 	GENERATED_BODY()
 public:
@@ -29,7 +29,7 @@ public:
 	 * @return The numerical ping (ms) to the session host.
 	 */
 	UFUNCTION(BlueprintPure, DisplayName = "Get Ping", Category = "NEXUS|Multiplayer", meta = (WorldContext = "WorldContextObject",
-		DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#ping"))
+		DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#ping"))
 	static float Ping(const UObject* WorldContextObject);
 
 	/**
@@ -40,7 +40,7 @@ public:
 	 * @return Was the player able to be kicked? true/false.
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Kick Player", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#kick-player"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#kick-player"))
 	static bool KickPlayer(UObject* WorldContextObject, APlayerState* PlayerState);
 
 	/**
@@ -50,7 +50,7 @@ public:
 	 * @return true/false if remote clients are found.
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has Remote Players", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-remote-players"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-remote-players"))
 	static bool HasRemotePlayers(UObject* WorldContextObject);
 
 	/**
@@ -60,7 +60,7 @@ public:
 	 * @return true/false if remote clients are found.
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has Remote Players ?", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-remote-players"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-remote-players"))
 	static bool HasRemotePlayersExec(UObject* WorldContextObject)
 	{
 		return HasRemotePlayers(WorldContextObject);
@@ -73,7 +73,7 @@ public:
 	 * @return true/false if only local clients are found.
 	 */	
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has Local Players", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-local-players-only"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-local-players-only"))
 	static bool HasLocalPlayersOnly(UObject* WorldContextObject)
 	{
 		return !HasRemotePlayers(WorldContextObject);
@@ -86,7 +86,7 @@ public:
 	 * @return true/false if only local clients are found.
 	 */		
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has Local Players ?", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-local-players-only"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-local-players-only"))
 	static bool HasLocalPlayersOnlyExec(UObject* WorldContextObject)
 	{
 		return HasLocalPlayersOnly(WorldContextObject);
@@ -99,7 +99,7 @@ public:
 	 * @return true/false if authority is found.
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has GameState Authority", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-gamestate-authority"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-gamestate-authority"))
 	static bool HasGameStateAuthority(UObject* WorldContextObject)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -116,7 +116,7 @@ public:
 	 * @return true/false if authority is found.
 	 */
 	UFUNCTION(BlueprintCallable, DisplayName = "Has GameState Authority ?", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-gamestate-authority"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-gamestate-authority"))
 	static bool HasGameStateAuthorityExec(UObject* WorldContextObject)
 	{
 		return HasGameStateAuthority(WorldContextObject);
@@ -129,7 +129,7 @@ public:
 	 * @return true/false if authority is found.
 	 */	
 	UFUNCTION(BlueprintCallable, BlueprintPure = false, DisplayName = "Has World Authority", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-world-authority"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-world-authority"))
 	static bool HasWorldAuthority(UObject* WorldContextObject)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -146,7 +146,7 @@ public:
 	 * @return true/false if authority is found.
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Has World Authority ?", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#has-world-authority"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#has-world-authority"))
 	static bool HasWorldAuthorityExec(UObject* WorldContextObject)
 	{
 		return HasWorldAuthority(WorldContextObject);
@@ -158,12 +158,12 @@ public:
 	 * @return The player's identifier.
 	 */
 	UFUNCTION(BlueprintCallable, DisplayName = "Get PlayerIdentifier", Category = "NEXUS|Multiplayer",
-			meta = (DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-playeridentifier"))
+			meta = (DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#get-playeridentifier"))
 	static int32 GetPlayerIdentifier(const APlayerController* PlayerController)
 	{
-		N_VALIDATE_RETURN(LogNexusMultiplayer, PlayerController, 0)
+		N_VALIDATE_RETURN(LogNexusCore, PlayerController, 0)
 		const APlayerState* PlayerState = PlayerController->GetPlayerState<APlayerState>();
-		N_VALIDATE_RETURN(LogNexusMultiplayer, PlayerState, 0)
+		N_VALIDATE_RETURN(LogNexusCore, PlayerState, 0)
 		return PlayerState->GetPlayerId();
 	}
 
@@ -173,7 +173,7 @@ public:
 	 * @return The player's identifier.
 	 */
 	UFUNCTION(BlueprintCallable, DisplayName = "Get First PlayerIdentifier", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-first-playeridentifier"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#get-first-playeridentifier"))
 	static int32 GetFirstPlayerIdentifier(UObject* WorldContextObject)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -190,7 +190,7 @@ public:
 	 * @return If found, APawn, or nullptr.
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Get Pawn From PlayerIdentifier", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-pawn-from-playeridentifier"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#get-pawn-from-playeridentifier"))
 	static APawn* GetPawnFromPlayerIdentifier(UObject* WorldContextObject, const int32 PlayerIdentifier)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -207,7 +207,7 @@ public:
 	 * @return If found, AActor, or nullptr.
 	 */
 	UFUNCTION(BlueprintCallable, DisplayName = "Get PlayerController From PlayerIdentifier", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-playercontroller-from-playeridentifier"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#get-playercontroller-from-playeridentifier"))
 	static AActor* GetPlayerControllerFromPlayerIdentifier(UObject* WorldContextObject, const int32 PlayerIdentifier)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -224,7 +224,7 @@ public:
 	 * @return If found, APlayerState, or nullptr.
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Get PlayerState From PlayerIdentifier", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#get-playerstate-from-playeridentifier"))
+		meta = (WorldContext = "WorldContextObject", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#get-playerstate-from-playeridentifier"))
 	static APlayerState* GetPlayerStateFromPlayerIdentifier(UObject* WorldContextObject, const int32 PlayerIdentifier)
 	{
 		if (const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject))
@@ -240,7 +240,7 @@ public:
 	 * @return true/false if the world is not operating in NM_Client mode.
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Is Server", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-server"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#is-server"))
 	static bool IsServer(UObject* WorldContextObject)
 	{
 		const UWorld* World = N_GET_WORLD_FROM_CONTEXT(WorldContextObject);
@@ -253,7 +253,7 @@ public:
 	 * @return true/false if the world is not operating in NM_Client mode.
 	 */	
 	UFUNCTION(BlueprintCallable, DisplayName = "Is Server ?", Category = "NEXUS|Multiplayer",
-		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/multiplayer/types/multiplayer-library/#is-server"))
+		meta = (WorldContext = "WorldContextObject", ExpandBoolAsExecs="ReturnValue", DocsURL="https://nexus-framework.com/docs/plugins/core/types/multiplayer-library/#is-server"))
 	static bool IsServerExec(UObject* WorldContextObject)
 	{
 		return IsServer(WorldContextObject);