+++ Samples/Shared/Source/NexusSharedSamples/Public/NSamplesPawn.h
@@ -59,7 +59,6 @@ private:
 		"/NexusActorPoolsSamples/",
 		"/NexusBlockoutSamples/",
 		"/NexusDynamicRefsSamples/"
-		"/NexusMultiplayerSamples/",
 		"/NexusPickerSamples/",
 		"/NexusProcGenSamples/",
 		"/NexusUISamples/",